requirements.txt
Flask==3.0.3
qrcode==7.4.2
Pillow==10.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-telegram-bot==21.2
pyzbar==0.1.9
supabase==2.3.5
schedule==1.2.0
